LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/constructor/021-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket 'foobar, foobar'

PASS WebSockets: Same sub protocol twice
